+++ packages/orq-rc/src/funcs/evalsAll.ts
@@ -11,7 +11,6 @@ import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
-import * as components from "../models/components/index.js";
 import {
   ConnectionError,
   InvalidRequestError,
@@ -19,6 +18,7 @@ import {
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
+import * as errors from "../models/errors/index.js";
 import { OrqError } from "../models/errors/orqerror.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
@@ -38,7 +38,8 @@ export function evalsAll(
   options?: RequestOptions,
 ): APIPromise<
   Result<
-    components.ListEvaluatorsResponse,
+    operations.GetEvalsResponseBody,
+    | errors.GetEvalsResponseBody
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -63,7 +64,8 @@ async function $do(
 ): Promise<
   [
     Result<
-      components.ListEvaluatorsResponse,
+      operations.GetEvalsResponseBody,
+      | errors.GetEvalsResponseBody
       | OrqError
       | ResponseValidationError
       | ConnectionError
@@ -150,8 +152,13 @@ async function $do(
   }
   const response = doResult.value;
 
+  const responseFields = {
+    HttpMeta: { Response: response, Request: req },
+  };
+
   const [result] = await M.match<
-    components.ListEvaluatorsResponse,
+    operations.GetEvalsResponseBody,
+    | errors.GetEvalsResponseBody
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -161,10 +168,11 @@ async function $do(
     | UnexpectedClientError
     | SDKValidationError
   >(
-    M.json(200, components.ListEvaluatorsResponse$inboundSchema),
+    M.json(200, operations.GetEvalsResponseBody$inboundSchema),
+    M.jsonErr(404, errors.GetEvalsResponseBody$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
-  )(response, req);
+  )(response, req, { extraFields: responseFields });
   if (!result.ok) {
     return [result, { status: "complete", request: req, response }];
   }