+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-16T15:04:51.742Z",
+    "2026-09-16T15:57:14.246Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {