+++ packages/orq-rc/RELEASES.md
@@ -12429,3 +12429,13 @@ Based on:
 - [typescript v4.15.0-rc.54] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.54] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.54 - packages/orq-rc
+
+## 2026-09-16 16:01:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.55] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.55] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.55 - packages/orq-rc